Change sendmail functions into classes that can be overriden from a plugin
