+++ CTFd/utils/email/providers/__init__.py
@@ -1,0 +1,4 @@
+class EmailProvider:
+    @staticmethod
+    def sendmail(addr, text, subject):
+        raise NotImplementedError
